requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.4.2
autopep8==1.5.1
certifi==2020.4.5.1
colorama==0.4.3
comtypes==1.1.7
distlib==0.3.0
Django==3.0.7
django-ckeditor==5.9.0
django-cleanup==5.0.0
django-crispy-forms==1.9.1
django-js-asset==1.2.2
filelock==3.0.12
importlib-metadata==1.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
pipenv==2018.11.26
pycodestyle==2.5.0
pylint==2.5.3
pynput==1.6.8
PyPDF2==1.26.0
pyttsx3==2.87
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
virtualenv==20.0.10
virtualenv-clone==0.5.4
wrapt==1.12.1
zipp==3.1.0
